ml/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["heka1024 <heka1024@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10, <3.13"
numpy = "^1.26.1"
pandas = "^2.1.2"
pytz = "^2023.3.post1"
matplotlib = "^3.8.1"
scikit-learn = "^1.3.2"
grpcio = "^1.59.2"
protobuf = "^4.25.0"
googleapis-common-protos = "^1.61.0"
geopandas = "^0.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
